refactor: prayerRead Dir 구조 변경 #25
- prayers로 통일
- notifications 도메인 생성
