+++ app/prayers/[id]/_components/index/PrayerReadBody.tsx
@@ -1,13 +1,13 @@
 import React, { useEffect } from "react";
 import { View, StyleSheet } from "react-native";
 import { RFValue } from "react-native-responsive-fontsize";
-import { PrayerContent } from "../types/response/prayerContent";
+import { PrayerContent } from "../../../../../src/domain/prayers/types/prayerContent";
 import TitleCard from "./TitleCard";
 import PrayerCardList from "./PrayerCardList";
 import EmptyState from "./EmptyState";
-import { useSelectedPrayerTitleStore } from "../../prayers/stores/useSelectedPrayerTitleStore";
-import { useSelectedRoomStore } from "../../rooms/stores/useSelectedRoomStore";
-import { usePrayerContentsQuery } from "../hooks/queries/usePrayerReadQuery";
+import { useSelectedPrayerTitleStore } from "../../../../../src/domain/prayers/stores/useSelectedPrayerTitleStore";
+import { useSelectedRoomStore } from "../../../../../src/domain/rooms/stores/useSelectedRoomStore";
+import { usePrayerContentsQuery } from "@/domain/prayers/hooks/queries/usePrayerQueries";
 
 interface PrayerReadBodyProps {
   onEdit?: (prayerContent: PrayerContent) => void;
