+++ app/prayers/[id]/_components/index/PrayerReadTop.tsx
@@ -3,10 +3,10 @@ import { StyleSheet } from "react-native";
 import { Appbar, useTheme } from "react-native-paper";
 import { useRouter } from "expo-router";
 import { RFValue } from "react-native-responsive-fontsize";
-import { color } from "../../../common/styles/color";
-import path from "../../../../src/common/constants/path";
-import { useSelectedRoomStore } from "../../rooms/stores/useSelectedRoomStore";
-import { useSelectedPrayerTitleStore } from "../../prayers/stores/useSelectedPrayerTitleStore";
+import { color } from "../../../../../src/common/styles/color";
+import path from "../../../../../src/common/constants/path";
+import { useSelectedRoomStore } from "../../../../../src/domain/rooms/stores/useSelectedRoomStore";
+import { useSelectedPrayerTitleStore } from "../../../../../src/domain/prayers/stores/useSelectedPrayerTitleStore";
 
 interface PrayerReadTopProps {}
