+++ app/rooms/_components/[id]/PrayerRoomBody.tsx
@@ -1,4 +1,4 @@
-import PrayerTitleList from "../../../../rooms/_components/[id]/prayerTitleList";
+import PrayerTitleList from "./prayerTitleList";
 
 export default function PrayerRoomBody(): JSX.Element {
   return <PrayerTitleList />;
