+++ app/rooms/_components/[id]/RoomMembersModal.tsx
@@ -2,8 +2,8 @@ import React, { useState } from "react";
 import { StyleSheet, Dimensions, View, ScrollView } from "react-native";
 import { Portal, Modal, Button, Text } from "react-native-paper";
 import { RFValue } from "react-native-responsive-fontsize";
-import { color } from "../../../../../src/common/styles/color";
-import { useSelectedRoomStore } from "../../../../../src/domain/rooms/stores/useSelectedRoomStore";
+import { color } from "../../../../src/common/styles/color";
+import { useSelectedRoomStore } from "../../../../src/domain/rooms/stores/useSelectedRoomStore";
 import { useRoomMembersQuery } from "@/domain/rooms/hooks/queries/roomQueries";
 
 const { width } = Dimensions.get("window");
