+++ src/common/constants/notificationType.ts
