+++ src/domain/notifications/hooks/mutations/useNotificationMutations.ts
@@ -2,16 +2,20 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import QUERY_KEYS from "../../../../common/constants/queryKeys";
 import { ApiError } from "../../../../common/apis/api";
 import { Alert } from "react-native";
-import type { NotificationRequest } from "../../types/request/notificationRequest";
-import prayerNotificationService from "./../../services/prayerNotificationService";
+import { NotifyPrayerCompletionRequest } from "../../../notifications/types/request/notifyPrayerCompletionRequest";
+import { notificationService } from "../../services/notificationService";
 
 // 기도 완료 알림
 export const usePrayerNotificationMutation = () => {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: ({ prayerTitleId, roomId, type }: NotificationRequest) => {
-      return prayerNotificationService.notifyPrayerCompletion({
+    mutationFn: ({
+      prayerTitleId,
+      roomId,
+      type,
+    }: NotifyPrayerCompletionRequest) => {
+      return notificationService.notifyPrayerCompletion({
         prayerTitleId,
         roomId,
         type,
