+++ src/domain/notifications/types/request/notifyPrayerCompletionRequest.ts
@@ -1,0 +1,7 @@
+import { NOTIFICATION_TYPE } from "../../../../common/constants/notificationType";
+
+export interface NotifyPrayerCompletionRequest {
+  prayerTitleId: number | null;
+  roomId: number | null;
+  type: NOTIFICATION_TYPE;
+}
