--- src/domain/prayerRead/types/request/notificationRequest.ts
@@ -1,7 +1,0 @@
-import { NOTIFICATION_TYPE } from "../../constants/notificationType";
-
-export interface NotificationRequest {
-  prayerTitleId: number | null;
-  roomId: number | null;
-  type: NOTIFICATION_TYPE;
-}
