+++ src/domain/prayers/types/response/fetchPrayerContentsResponse.ts
@@ -1,0 +1,5 @@
+import { PrayerContent } from "../prayerContent";
+
+export interface FetchPrayerContentsResponse {
+  prayerContents: PrayerContent[];
+}
